tools/requirements.txt
Pillow==10.3.0
questionary==2.0.1
